Added SimpleDeploymentContext and have ConfigurationBasedDeploymentContext extend it.
